+++ packages/snap-account-service/src/index.ts
@@ -1,6 +1,11 @@
 export { SnapAccountService } from './SnapAccountService';
 export type {
   SnapAccountServiceActions,
+  SnapAccountServiceConfig,
   SnapAccountServiceEvents,
   SnapAccountServiceMessenger,
+  SnapAccountServiceOptions,
 } from './SnapAccountService';
+export type { SnapAccountServiceEnsureReadyAction } from './SnapAccountService-method-action-types';
+export { SnapPlatformWatcher } from './SnapPlatformWatcher';
+export type { SnapPlatformWatcherConfig } from './SnapPlatformWatcher';